# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 관계온도 API
  version: 1.1.0
  summary: 카카오톡 대화 기반 관계 건강도 분석 및 AI 상담 API
  description: |
    관계 등록, 카카오톡 대화 업로드, 비동기 PRQC 분석, 주간 대시보드,
    인물별 리포트와 근거 기반 AI 상담을 제공하는 MVP API 계약이다.

    브라우저 세션은 HttpOnly 쿠키를 사용한다. 상태 변경 요청은 세션 쿠키와
    X-CSRF-Token을 모두 요구한다. 점수와 AI 답변은 진단이 아닌 참고 정보다.

    프론트엔드-백엔드 분석은 비동기 Job이지만 Backend Worker-AI 서버 호출은
    MVP에서 동기 방식이다. Job의 stage와 progress는 백엔드 오케스트레이션
    기준의 예상값이다. AI 서버는 PRQC 6요소와 근거를 반환하고 백엔드는
    버전된 관계 유형별 정책으로 canonical overall.score를 계산한다.
servers:
  - url: /api/v1
    description: 현재 호스트의 v1 API
...
security:
  - cookieAuth: []
paths:
...
  /relationships/{relationshipId}/check-ins:
    parameters:
      - $ref: "#/components/parameters/RelationshipId"
    post:
      tags: [Check-ins]
      operationId: createCheckIn
      summary: 1~7점 주관적 체크인 저장
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckInRequest"
      responses:
        "201":
          description: 체크인 저장 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: rt_session
      description: HttpOnly Secure 세션 쿠키
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: /users/me에서 받은 CSRF 토큰
  parameters:
    RelationshipId:
      name: relationshipId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Forbidden:
      description: CSRF 검증 실패 또는 접근 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    UnprocessableEntity:
      description: 의미 검증 실패
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Uuid:
      type: string
      format: uuid
    DateTime:
      type: string
      format: date-time
...
    ErrorField:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ErrorField"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
...
    CheckInQuestionCode:
      type: string
      enum: [RELATIONSHIP_FEELING, CONVERSATION_COMFORT]
    CheckInAnswer:
      type: object
      additionalProperties: false
      required: [questionCode, score]
      properties:
        questionCode:
          $ref: "#/components/schemas/CheckInQuestionCode"
        score:
          type: integer
          minimum: 1
          maximum: 7
    CreateCheckInRequest:
      type: object
      additionalProperties: false
      required: [answers]
      properties:
        answers:
          type: array
          minItems: 2
          maxItems: 2
          items:
            $ref: "#/components/schemas/CheckInAnswer"
    CheckIn:
      type: object
      additionalProperties: false
      required: [id, relationshipId, answers, weekStart, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        relationshipId:
          $ref: "#/components/schemas/Uuid"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/CheckInAnswer"
        weekStart:
          type: string
          format: date
        createdAt:
          $ref: "#/components/schemas/DateTime"
    CheckInEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CheckIn"
...